fix: Unable to cast object of type 'System.Int64' to type 'System.Int32 (Comment id : int ==> long)

+++ MB.Application.Contracts/Comment/ICommentApplication.cs
@@ -6,8 +6,8 @@
 
          List<CommentViewModel> GetList();
 
-         void Confirm(int id);
+         void Confirm(long id);
 
-         void Cancel(int id);
+         void Cancel(long id);
     }
 }
